requirements/prod.txt
Django==5.1.3
python-dotenv==1.0.1
pillow==11.0.0
django-ckeditor==6.7.2
